refactor(web): extract InstallButton from OverviewPage
The page held the confirmation popup, the button state (danger variant,
disabled reasons, label) and the helper texts explaining why installation
cannot start yet. All of it now lives in a single component, leaving the
page to deal with layout only.

No visible change.
